+++ resources/lang/en/messages.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 return [
 
-    'already_authenticated' => 'You are already authenticated!',
     'import_complete' => 'Congrats! Import process completed.',
     'resource_saved' => 'Congrats! Requested :resource [:identifier] saved successfully.',
     'resource_deleted' => 'Done! Requested :resource [:identifier] deleted successfully.',
@@ -12,7 +11,6 @@ return [
     'resource_not_found' => 'Sorry! Requested :resource [:identifier] not found!',
     'delete_confirmation_title' => 'Delete Confirmation',
     'delete_confirmation_body' => 'Are you sure you want to delete :resource [:identifier]?',
-    'session_required' => 'You must login first!',
     'token_mismatch' => 'The page has expired due to inactivity. Please refresh and try again.',
     'invalid_phone' => 'Invalid phone number!',
     'invalid_indexing_rule' => 'An indexing rule needs to return a boolean or a string.',