+++ resources/views/adminarea/errors/419.blade.php
@@ -1,4 +1,4 @@
-@extends('cortex/foundation::common.errors.layout')
+@extends('cortex/foundation::adminarea.layouts.error')
 
 @section('title', __('Page Expired'))
 @section('code', '419')