+++ resources/views/frontarea/errors/403.blade.php
@@ -1,0 +1,5 @@
+@extends('cortex/foundation::frontarea.layouts.error')
+
+@section('title', __('Forbidden'))
+@section('code', '403')
+@section('message', __('Forbidden'))