+++ resources/views/frontarea/errors/503.blade.php
@@ -1,0 +1,5 @@
+@extends('cortex/foundation::frontarea.layouts.error')
+
+@section('title', __('Service Unavailable'))
+@section('code', '503')
+@section('message', __('Service Unavailable'))