+++ resources/views/managerarea/errors/419.blade.php
@@ -1,0 +1,5 @@
+@extends('cortex/foundation::managerarea.layouts.error')
+
+@section('title', __('Page Expired'))
+@section('code', '419')
+@section('message', __('Page Expired'))